requirements.txt
cython
numpy
albumentations
imagecorruptions
mmcv-full==1.3.15
matplotlib
terminaltables
six
pycocotools
pillow==6.2.2
timm
